+++ src/main/java/nextstep/global/exception/DuplicateEntityException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.global.exception;
 
 public class DuplicateEntityException extends RuntimeException {
 }
