+++ src/main/java/nextstep/global/exception/UnauthenticatedException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.global.exception;
 
 public class UnauthenticatedException extends RuntimeException {
 }
